fix forget password fix env var
